Fixing IVC (#79)
* Fixing IVC

* clippy

* fixup verifier cost (remove CI)

* Fixup compute costs (from ivc recombining fixed base)

* fixup : remove dead code

* Fixup cost estimation of IVC example

* remove old print

* fixups

* Sorting msm in circuit_stats for easier comparison

* fixup plinth

+++ src/plutus_gen/stats/chips/primitives/curve/ecc/weierstrass/mod.rs
@@ -65,8 +65,7 @@ pub(crate) trait WeierstrassChipTrait<P: WeierstrassEmulationParams>:
     FieldChipTrait<P>
 {
     fn advice() -> Vec<Column> {
-        let ecc_len =
-            P::NB_LIMBS as usize + std::cmp::max(P::NB_LIMBS as usize, 2 + P::moduli().len()) + 1;
+        let ecc_len = P::NB_LIMBS + std::cmp::max(P::NB_LIMBS, 2 + P::moduli().len()) + 1;
         let lambda2_advices = <Lambda2Chip as WierstrassOpChipTrait<P>>::advice();
         let oncurve_advices = <OnCurveChip as WierstrassOpChipTrait<P>>::advice();
         let slope_advices = <SlopeChip as WierstrassOpChipTrait<P>>::advice();
